Move sidebar item models to seperate files, common folder
