+++ Files/DataModels/SidebarPinnedModel.cs
@@ -1,5 +1,6 @@
 ﻿using Files.Common;
 using Files.Controllers;
+using Files.DataModels.NavigationControlItems;
 using Files.Filesystem;
 using Files.Helpers;
 using Files.UserControls;
