+++ Files/Filesystem/LibraryManager.cs
@@ -1,4 +1,5 @@
 ﻿using Files.Common;
+using Files.DataModels.NavigationControlItems;
 using Files.Extensions;
 using Files.Helpers;
 using Files.UserControls;
@@ -13,7 +14,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Windows.ApplicationModel.Core;
 using Windows.UI.Core;
-using Windows.UI.Xaml.Media;
 
 namespace Files.Filesystem
 {
