+++ Files/ViewModels/Properties/DriveProperties.cs
@@ -1,4 +1,4 @@
-﻿using Files.Filesystem;
+﻿using Files.DataModels.NavigationControlItems;
 using Microsoft.Toolkit.Uwp;
 using System;
 using System.Threading.Tasks;
