+++ Files/Views/MainPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using Files.Filesystem;
+﻿using Files.DataModels.NavigationControlItems;
+using Files.Filesystem;
 using Files.Helpers;
 using Files.UserControls;
 using Files.UserControls.MultitaskingControl;
